+++ src/gufo/acme/clients/__init__.py
@@ -11,5 +11,6 @@ The following clients are provided out-of-the-box:
 
 * [base][gufo.acme.clients.base] - Base class
 * [dav][gufo.acme.clients.dav] - http-01 WebDAV fulfillment.
+* [powerdns][gufo.acme.clients.powerdns] - dns-01 PowerDNS fulfillment.
 * [web][gufo.acme.clients.web] - http-01 static file fulfillment.
 """